feat(scheduler) codify scheduler component
To keep things simple, we want to use Heroku Scheduler. Unfortunately,
that scheduler has very limited control over actual scheduling. This
codifies Wednesday as the day we wish to run the Team Train.

On Heroku Scheduler, there will be 4 corresponding jobs that run daily
at the proper time, but this will catch and exit if it isn't Wednesday.